+++ src/main/java/com/kakaotech/team14backend/common/FileUtils.java
@@ -1,14 +1,15 @@
 package com.kakaotech.team14backend.common;
 
-import com.kakaotech.team14backend.exception.ExtentionNotAllowedException;
+import com.kakaotech.team14backend.image.exception.ImageIOException;
+import com.kakaotech.team14backend.post.exception.ExtentionNotAllowedException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
+
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.UUID;
-import com.kakaotech.team14backend.exception.ImageIOException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
 
 @Component
 @Slf4j
